Redefining declared variables to support dynamic expressions
